refactor(#668): move LegalPrefsStore, NotificationPrefsStore, SettingsStore to Preferences/
These three stores are app-wide preference concerns, not runner domain
logic. Moving them out of Runner/ into the new Preferences/ subfolder
corrects the misleading grouping. No content changes.

- Add Sources/RunnerBar/Preferences/LegalPrefsStore.swift
- Add Sources/RunnerBar/Preferences/NotificationPrefsStore.swift
- Add Sources/RunnerBar/Preferences/SettingsStore.swift
- Stub old Runner/ files with redirect comments (safe to delete)

Fixes #668, part of #654
